code refactoring codingChallenge and article and api integration
