--- src/app/admin/courses/[courseId]/module/_components/Article.tsx
@@ -1,8 +1,0 @@
-import React, { useState } from 'react'
-
-function Article({ content }: { content: Object }) {
-    const [articleContent, setArticleContent] = useState([])
-    return <div>Article</div>
-}
-
-export default Article
