+++ src/app/admin/courses/[courseId]/module/_components/Article/Tiptap.tsx
@@ -9,7 +9,6 @@ import {
     Italic,
     List,
     ListOrdered,
-    Quote,
     Minus,
     Undo,
     Redo,
